+++ src/Pages/WorkoutPage/WorkoutTypes/Run/HeartRateDeviceModal.js
@@ -4,6 +4,7 @@ import {
   Modal,
   Pressable,
   ScrollView,
+  StyleSheet,
   TouchableOpacity,
   View,
 } from "react-native";